Move BlockWithHead to separate file
